// developershub-frontend/developers/src/app/model/developer.ts
export class Developer {
  id: number;
  fristName: string;
  lastName: string;
  age: number;
  email: string;
  phoneNumber: string;
  jobTitle: string;
  companyWorkFor: string;
  typeOfDeveloper: string;
  programmingLanguages: string
  isEmployed: boolean;
  address: [];
  companyEmployedTo: string;
  recruiters: [];

}